refactor(client): move per-lake file hooks from the wizard module to dataLakes
